Add App.jsx wrapper
