+++ src/index.jsx
@@ -1,34 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { IntlProvider } from 'react-intl';
-import { QueryClientProvider } from 'react-query';
 
-import { loadTranslations } from './loadTranslations';
-import { isValidConfig } from './loginServices';
-import createReactQueryClient from './createReactQueryClient';
-import Router from './Router';
-import ConfigError from './ConfigError';
-import configuration from '../config.yaml';
-
-const config = configuration.config || {};
-const branding = configuration.branding || {};
-const location = globalThis.location;
-const locale = config.locale || navigator.language || 'en-US';
-const translations = await loadTranslations(locale);
-const reactQueryClient = createReactQueryClient();
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-const Component = isValidConfig(config) ?
-  <Router branding={branding} config={config} location={location} />
-  :
-  <ConfigError branding={branding} config={config} />
 
 root.render(
   <React.StrictMode>
-    <QueryClientProvider client={reactQueryClient}>
-      <IntlProvider locale={locale} messages={translations} >
-        {Component}
-      </IntlProvider>
-    </QueryClientProvider>
+    <App />
   </React.StrictMode>
 );
